Refactor: move getRandomName to pageComponent folder
