+++ src/frontend/src/pages/FlowPage/components/PageComponent/utils/get-random-name.tsx
@@ -1,0 +1,35 @@
+import { ADJECTIVES, NOUNS } from "../../../../../flow_constants";
+import { getRandomElement } from "../../../../../utils/reactflowUtils";
+import { toTitleCase } from "../../../../../utils/utils";
+
+export default function getRandomName(
+    retry: number = 0,
+    noSpace: boolean = false,
+    maxRetries: number = 3,
+  ): string {
+    const left: string[] = ADJECTIVES;
+    const right: string[] = NOUNS;
+  
+    const lv = getRandomElement(left);
+    const rv = getRandomElement(right);
+  
+    // Condition to avoid "boring wozniak"
+    if (lv === "boring" && rv === "wozniak") {
+      if (retry < maxRetries) {
+        return getRandomName(retry + 1, noSpace, maxRetries);
+      } else {
+        console.warn("Max retries reached, returning as is");
+      }
+    }
+  
+    // Append a suffix if retrying and noSpace is true
+    if (retry > 0 && noSpace) {
+      const retrySuffix = Math.floor(Math.random() * 10);
+      return `${lv}_${rv}${retrySuffix}`;
+    }
+  
+    // Construct the final name
+    let final_name = noSpace ? `${lv}_${rv}` : `${lv} ${rv}`;
+    // Return title case final name
+    return toTitleCase(final_name);
+  }
